apps/open-source/token-server/requirements.txt
fastapi
uvicorn
python-dotenv
livekit-api